feat: 开发归物页面4

+++ fortuneboot-domain/src/main/java/com/fortuneboot/domain/command/fortune/FortuneGoodsKeeperAddCommand.java
@@ -90,7 +90,7 @@ public class FortuneGoodsKeeperAddCommand {
     /**
      * 出售价格
      */
-    private Integer soldPrice;
+    private BigDecimal soldPrice;
 
     /**
      * 备注
